+++ burnxp-candle/src/main.rs
@@ -1,0 +1,5 @@
+use burnxp_core::run;
+
+fn main() {
+    run()
+}