+++ burnxp-core/src/cli.rs
@@ -1,8 +1,5 @@
-use burn::{
-    backend::{libtorch::LibTorchDevice, Autodiff, LibTorch},
-    optim::AdamConfig,
-};
-use burnxp::{predict, train, ModelConfig, Output, PredictConfig, ResNetType, TrainingConfig};
+use crate::{predict, train, ModelConfig, Output, PredictConfig, ResNetType, TrainingConfig};
+use burn::{backend::Autodiff, optim::AdamConfig};
 use clap::{CommandFactory as _, Parser, Subcommand};
 use clap_complete::{generate, Shell};
 use std::path::PathBuf;
@@ -86,14 +83,23 @@ enum SubCmd {
     },
 }
 
-type MyBackend = LibTorch<f32, i8>;
+#[cfg(feature = "tch")]
+type MyBackend = burn::backend::LibTorch<f32, i8>;
+#[cfg(feature = "candle")]
+type MyBackend = burn::backend::Candle<f32, u8>;
+
 type MyAutodiffBackend = Autodiff<MyBackend>;
 
-fn main() {
-    #[cfg(target_os = "macos")]
-    let device = LibTorchDevice::Mps;
-    #[cfg(not(target_os = "macos"))]
-    let device = LibTorchDevice::Cuda(0);
+pub fn run() {
+    #[cfg(all(feature = "tch", target_os = "macos"))]
+    let device = burn::backend::libtorch::LibTorchDevice::Mps;
+    #[cfg(all(feature = "tch", not(target_os = "macos")))]
+    let device = burn::backend::libtorch::LibTorchDevice::Cuda(0);
+
+    #[cfg(all(feature = "candle", target_os = "macos"))]
+    let device = burn::backend::candle::CandleDevice::Metal(0);
+    #[cfg(all(feature = "candle", not(target_os = "macos")))]
+    let device = burn::backend::candle::CandleDevice::Cuda(0);
 
     let args = Cli::parse();
     match args.subcmd {