+++ burnxp-core/src/lib.rs
@@ -1,8 +1,12 @@
+#![cfg(any(feature = "tch", feature = "candle"))]
+
+mod cli;
 mod data;
 mod model;
 mod predict;
 mod train;
 
+pub use cli::run;
 pub use model::{ModelConfig, ResNetType};
 pub use predict::{predict, Output, PredictConfig};
 pub use train::{train, TrainingConfig};